+++ com.unity.render-pipelines.high-definition/CHANGELOG.md
@@ -95,6 +95,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Added support for stack lit for ray tracing effects.
 - Added support for hair for ray tracing effects.
 - Added support for alpha to coverage for HDRP shaders and shader graph
+- Added support for Quality Levels to Subsurface Scattering.
 
 ### Fixed
 - Fix when rescale probe all direction below zero (1219246)